+++ Metadata.cs
@@ -4,6 +4,6 @@
     {
         public const string GUID = "ShaosilGaming.PeakGeneralImprovements";
         public const string PLUGIN_NAME = "PeakGeneralImprovements";
-        public const string VERSION = "1.0.2";
+        public const string VERSION = "1.0.3";
     }
 }