+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+### v1.0.3 - Hotfix and new option
+* Tweaked the auto dismount behavior to more accurately dismount when on certain sides of vines.
+* Added `SpawnMissingPropsOnLateJoins` as a host-only option (defaults to true) that will automatically spawn missing props like marshmallows when a player joins in the middle of a game.
+
 ### v1.0.2 - Bugfixes and improvements
 * Fixed the `DisableFogTimer` setting showing up as `FogSetting` in the config.
 * Fixed a bug where errors would spam the log after returning to airport if `CampfiresPreventHunger` was set to true.
-* Added an option (defaults to `AllowClimbing`) to change the behavior of ropes/vines/chains. Options:
+* Added `RopeVineChainBehavior` as an option (defaults to `AllowClimbing`) to change the behavior of ropes/vines/chains. Options:
 	* `Vanilla` - Unchanged gameplay
 	* `AllowClimbing` - Allows player to climb on to another surface without first jumping
 	* `AutoDismount`- Will also automatically get off and attempt to start climbing at the end